structured cases into separate classes, moved preparation for runs into a separate base class
