+++ src/test/java/BaseTest.java
@@ -1,0 +1,24 @@
+
+import org.junit.After;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+import org.openqa.selenium.firefox.FirefoxDriver;
+import org.openqa.selenium.firefox.FirefoxOptions;
+
+public class BaseTest {
+    protected WebDriver driver;
+
+    public void setUpBrowser(String browser) {
+        if (browser.equals("chrome")) {
+            ChromeOptions options = new ChromeOptions();
+            options.addArguments("--no-sandbox", "--headless", "--disable-dev-shm-usage");
+            driver = new ChromeDriver(options);
+            driver.get("https://qa-scooter.praktikum-services.ru");
+        } else if (browser.equals("firefox")) {
+            FirefoxOptions options = new FirefoxOptions();
+            options.addArguments("--no-sandbox", "--headless", "--disable-dev-shm-usage");
+            driver = new FirefoxDriver(options);
+        }
+    }
+}
